Fail CMake setup if trying to build with Python 2
Summary:
Python 2 has reached end-of-life and is no longer supported by PyTorch.
To spare users from a long, doomed build when trying to use PyTorch with
Python 2, detect this case early and fail with a clear message.  This
commit covers CMake setup.

Test Plan:
Attempted to build PyTorch with Python 2 and saw a clear error *quickly*.

ghstack-source-id: c31d0df93cae85ec07fbde0a6b29033360803aec
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35612