Remove (most) Python 2 support from C++ code
Summary:
Python 2 has reached end-of-life and is no longer supported by PyTorch.
Now we can clean up a lot of cruft that we put in place to support it.
These changes were all done manually, and I skipped anything that seemed
like it would take more than a few seconds, so I think it makes sense to
review it manually as well.

Test Plan:
CI

ghstack-source-id: 728daaa086058b217ded56f0bc7e2b4a5d7fb3b7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35614