Fail setup.py if trying to set up with Python 2
Summary:
Python 2 has reached end-of-life and is no longer supported by PyTorch.
To spare users from a long, doomed setup when trying to use PyTorch with
Python 2, detect this case early and fail with a clear message.  This
commit covers setup.py.

Test Plan:
Attempted to build PyTorch with Python 2 and saw a clear error *quickly*.

ghstack-source-id: 07b6cd1f9af3143af782f471ad036a64dc5d7d2d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35613