qactivation_benchmarks: small bug fix
Summary:

Changes relu and relu6 to point to the functional implementations here.
The previous behavior tested the time to create the module, but didn't actually run the
function (I noticed this when adding the new input sizes and seeing
the measured time not change).

Test Plan:

run the benchmark, the time now changes as expected with input size for
these.

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 9a1ac41d350b90a1057f3c4a6cd87dbbd0066289
Pull Request resolved: https://github.com/pytorch/pytorch/pull/35731