Catch exception in distributed engine callbacks.
Callbacks registered with the autograd engine Future in the
distributed engine have a non-trivial amount of business logic. Its entirely
possible that we throw exceptions in these callbacks resulting in those not
being propagated back to the client (since the appropriate future was not
marked as completed).

In this PR, I've added appropriate try-catch blocks to ensure we always mark
the appropriate Future with an error.

Differential Revision: [D20885521](https://our.internmc.facebook.com/intern/diff/D20885521/)

ghstack-source-id: 101646113
Pull Request resolved: https://github.com/pytorch/pytorch/pull/36118